op-cancel-appointment - move encrypted-bark-report into its own model file
