+++ src/lib/bark/bark-models.ts
@@ -1,5 +1,2 @@
-import { BarkReport } from "./models/bark-report";
 
-export type EncryptedBarkReport = Omit<BarkReport, "ineligibilityReason"> & {
-  encryptedIneligibilityReason: string;
-};
+
