+++ src/lib/bark/mappers/to-bark-report.ts
@@ -1,5 +1,5 @@
 import { BarkContext } from "../bark-context";
-import { EncryptedBarkReport } from "../bark-models";
+import { EncryptedBarkReport } from "../models/encrypted-bark-report";
 import { BarkReport } from "../models/bark-report";
 
 export async function toBarkReport(
