+++ src/lib/bark/models/encrypted-bark-report.ts
@@ -1,0 +1,6 @@
+import { BarkReport } from "./bark-report";
+
+
+export type EncryptedBarkReport = Omit<BarkReport, "ineligibilityReason"> & {
+  encryptedIneligibilityReason: string;
+};
