+++ src/lib/bark/operations/op-fetch-report.ts
@@ -1,4 +1,4 @@
-import { EncryptedBarkReport } from "@/lib/bark/bark-models";
+import { EncryptedBarkReport } from "../models/encrypted-bark-report";
 import { BarkReport } from "../models/bark-report";
 import { CODE } from "@/lib/utilities/bark-code";
 import { Err, Ok, Result } from "@/lib/utilities/result";
